FaaS_GCP/Batch_with_scheduler/delete_messages_expired/.requirements.txt
functions-framework==3.*
motor
python-dotenv
pymongo